Finished init_finger_table
Potential bug:
Joining node does some RPCs to fill its finger table. However, for example, if
only the bootstrap node is in the network and the new joining node asks it for
an ID it does not know, the bootstrap node currently refers itself in its
finger table. Therefore, it returns a wrong responsibility. Maybe, this is
fixed later automatically. BUT check twice!